Renovated StepOne.js
